+++ src/types/observablevalue.ts
@@ -27,14 +27,28 @@ import {
 } from "../internal"
 
 export interface IValueWillChange<T> {
-    object: any
+    object: IObservableValue<T>
     type: "update"
     newValue: T
 }
 
-export interface IValueDidChange<T> extends IValueWillChange<T> {
-    oldValue: T | undefined
+export type IValueDidChange<T = any> = {
+    type: "update"
+    observableKind: "value"
+    object: IObservableValue<T>
+    debugObjectName: string
+    newValue: unknown
+    oldValue: unknown
 }
+export type IBoxDidChange<T = any> =
+    | {
+          type: "create"
+          observableKind: "value"
+          object: IObservableValue<T>
+          debugObjectName: string
+          newValue: unknown
+      }
+    | IValueDidChange<T>
 
 export interface IObservableValue<T> {
     get(): T
@@ -64,7 +78,13 @@ export class ObservableValue<T> extends Atom
         this.value_ = enhancer(value, undefined, name_)
         if (__DEV__ && notifySpy && isSpyEnabled()) {
             // only notify spy if this is a stand-alone observable
-            spyReport({ type: CREATE, name: this.name_, newValue: "" + this.value_ })
+            spyReport({
+                type: CREATE,
+                object: this,
+                observableKind: "value",
+                debugObjectName: this.name_,
+                newValue: "" + this.value_
+            })
         }
     }
 
@@ -81,7 +101,9 @@ export class ObservableValue<T> extends Atom
             if (__DEV__ && notifySpy) {
                 spyReportStart({
                     type: UPDATE,
-                    name: this.name_,
+                    object: this,
+                    observableKind: "value",
+                    debugObjectName: this.name_,
                     newValue,
                     oldValue
                 })
@@ -133,6 +155,8 @@ export class ObservableValue<T> extends Atom
     observe_(listener: (change: IValueDidChange<T>) => void, fireImmediately?: boolean): Lambda {
         if (fireImmediately)
             listener({
+                observableKind: "value",
+                debugObjectName: this.name_,
                 object: this,
                 type: UPDATE,
                 newValue: this.value_,